--- graph_weather/models/analysis.py
@@ -1,1 +1,0 @@
-"""Model for creating the analysis outputs for initializing NWPs"""
