+++ graph_weather/models/layers/normalization.py
@@ -1,54 +1,87 @@
-import torch
+from torch import float, max, ones, sqrt, sum, tensor, zeros
+from torch.nn import Module, Parameter
 
 
-class Normalizer(snt.AbstractModule):
-    """Feature normalizer that accumulates statistics online."""
+# to do: convert to a pytorch-compatible layer, modify MeshGraphNets.py (see BatchNorm for an example)
+# and use incremental update formulas for mean/std and/or exponential decay averaging
+class Normalizer(Module):
+    def __init__(self, size, max_accumulations=10 ** 7, epsilon=1e-8, device=None):
 
-    def __init__(self, size, max_accumulations=10 ** 6, std_epsilon=1e-8, name="Normalizer"):
-        super(Normalizer, self).__init__(name=name)
-        self._max_accumulations = max_accumulations
-        self._std_epsilon = std_epsilon
-        with self._enter_variable_scope():
-            self._acc_count = tf.Variable(0, dtype=tf.float32, trainable=False)
-            self._num_accumulations = tf.Variable(0, dtype=tf.float32, trainable=False)
-            self._acc_sum = tf.Variable(tf.zeros(size, tf.float32), trainable=False)
-            self._acc_sum_squared = tf.Variable(tf.zeros(size, tf.float32), trainable=False)
+        """
+        Online normalization module
 
-    def _build(self, batched_data, accumulate=True):
-        """Normalizes input data and accumulates statistics."""
-        update_op = tf.no_op()
-        if accumulate:
-            # stop accumulating after a million updates, to prevent accuracy issues
-            update_op = tf.cond(
-                self._num_accumulations < self._max_accumulations,
-                lambda: self._accumulate(batched_data),
-                tf.no_op,
+        size: feature dimension
+        max_accumulation: maximum number of batches
+        epsilon: std cutoff for constant variable
+        device: pytorch device
+        """
+
+        super(Normalizer, self).__init__()
+
+        self.max_accumulations = max_accumulations
+        self.epsilon = epsilon
+
+        # self.register_buffer('acc_count', tensor(0, dtype=float, device=device))
+        # self.register_buffer('num_accumulations', tensor(0, dtype=float, device=device))
+        # self.register_buffer('acc_sum', zeros(size, dtype=float, device=device))
+        # self.register_buffer('acc_sum_squared', zeros(size, dtype=float, device=device))
+
+        self.register_buffer("acc_count", tensor(1.0, dtype=float, device=device))
+        self.register_buffer("num_accumulations", tensor(1.0, dtype=float, device=device))
+        self.register_buffer("acc_sum", ones(size, dtype=float, device=device))
+        self.register_buffer("acc_sum_squared", ones(size, dtype=float, device=device))
+
+    def forward(self, batched_data, accumulate=True):
+        """
+        Updates mean/standard deviation and normalizes input data
+
+        batched_data: batch of data
+        accumulate: if True, update accumulation statistics
+        """
+        if accumulate and self.num_accumulations < self.max_accumulations:
+            self._accumulate(batched_data)
+
+        return (batched_data - self._mean().to(batched_data.device)) / self._std().to(
+            batched_data.device
             )
-        with tf.control_dependencies([update_op]):
-            return (batched_data - self._mean()) / self._std_with_epsilon()
 
-    @snt.reuse_variables
     def inverse(self, normalized_batch_data):
-        """Inverse transformation of the normalizer."""
-        return normalized_batch_data * self._std_with_epsilon() + self._mean()
+        """
+        Unnormalizes input data
+        """
+
+        return normalized_batch_data * self._std().to(
+            normalized_batch_data.device
+            ) + self._mean().to(normalized_batch_data.device)
 
     def _accumulate(self, batched_data):
-        """Function to perform the accumulation of the batch_data statistics."""
-        count = tf.cast(tf.shape(batched_data)[0], tf.float32)
-        data_sum = tf.reduce_sum(batched_data, axis=0)
-        squared_data_sum = tf.reduce_sum(batched_data ** 2, axis=0)
-        return tf.group(
-            tf.assign_add(self._acc_sum, data_sum),
-            tf.assign_add(self._acc_sum_squared, squared_data_sum),
-            tf.assign_add(self._acc_count, count),
-            tf.assign_add(self._num_accumulations, 1.0),
-        )
+        """
+        Accumulates statistics for mean/standard deviation computation
+        """
+        count = tensor(batched_data.shape[0]).float()
+        data_sum = sum(batched_data, dim=0)
+        squared_data_sum = sum(batched_data ** 2, dim=0)
+
+        self.acc_sum += data_sum.to(self.acc_sum.device)
+        self.acc_sum_squared += squared_data_sum.to(self.acc_sum_squared.device)
+        self.acc_count += count.to(self.acc_count.device)
+        self.num_accumulations += 1
 
     def _mean(self):
-        safe_count = tf.maximum(self._acc_count, 1.0)
-        return self._acc_sum / safe_count
+        """
+        Returns accumulated mean
+        """
+        safe_count = max(self.acc_count, tensor(1.0).float())
 
-    def _std_with_epsilon(self):
-        safe_count = tf.maximum(self._acc_count, 1.0)
-        std = tf.sqrt(self._acc_sum_squared / safe_count - self._mean() ** 2)
-        return tf.math.maximum(std, self._std_epsilon)
+        return self.acc_sum / safe_count
+
+    def _std(self):
+        """
+        Returns accumulated standard deviation
+        """
+        safe_count = max(self.acc_count, tensor(1.0).float())
+        std = sqrt(self.acc_sum_squared / safe_count - self._mean() ** 2)
+
+        std[std < self.epsilon] = 1.0
+
+        return std
